LGDDumper's entity- and tag filters now based on a filter-file (rather than a direct SQL query): Therefore dump and live-sync can use the same config (simplifies consisteny).
Fixed a bug where ontology resources corresponding to tags of the form ('value'=null) lacked a definition.

Retrieving instances of classes corresponding to tags of the form ('value'=null) now retrieves all 'subclasses' - to be more accurate: All entities with tags with an corresponding key -regardless of the values  - are fetched.


git-svn-id: http://linkedgeodata.googlecode.com/svn/trunk@143 00633da9-98c1-349d-95eb-89b68c819386